Added PSR7 middleware (and deprecated existing)
