+++ src/Middleware/AuthorizationServerMiddleware.php
@@ -9,47 +9,9 @@
 
 namespace League\OAuth2\Server\Middleware;
 
-use League\OAuth2\Server\AuthorizationServer;
-use League\OAuth2\Server\Exception\OAuthServerException;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
-
-class AuthorizationServerMiddleware
+/**
+ * @deprecated Use the Psr7AuthorizationServerMiddleware or SymfonyAuthorizationServerMiddleware
+ */
+class AuthorizationServerMiddleware extends PSR7AuthorizationServerMiddleware
 {
-    /**
-     * @var AuthorizationServer
-     */
-    private $server;
-
-    /**
-     * @param AuthorizationServer $server
-     */
-    public function __construct(AuthorizationServer $server)
-    {
-        $this->server = $server;
-    }
-
-    /**
-     * @param ServerRequestInterface $request
-     * @param ResponseInterface      $response
-     * @param callable               $next
-     *
-     * @return ResponseInterface
-     */
-    public function __invoke(ServerRequestInterface $request, ResponseInterface $response, callable $next)
-    {
-        try {
-            $response = $this->server->respondToAccessTokenRequest($request, $response);
-        } catch (OAuthServerException $exception) {
-            return $exception->generateHttpResponse($response);
-            // @codeCoverageIgnoreStart
-        } catch (\Exception $exception) {
-            return (new OAuthServerException($exception->getMessage(), 0, 'unknown_error', 500))
-                ->generateHttpResponse($response);
-            // @codeCoverageIgnoreEnd
-        }
-
-        // Pass the request and response on to the next responder in the chain
-        return $next($request, $response);
-    }
 }
