+++ src/Middleware/ResourceServerMiddleware.php
@@ -9,47 +9,9 @@
 
 namespace League\OAuth2\Server\Middleware;
 
-use League\OAuth2\Server\Exception\OAuthServerException;
-use League\OAuth2\Server\ResourceServer;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
-
-class ResourceServerMiddleware
+/**
+ * @deprecated Use the Psr7ResourceServerMiddleware or SymfonyResourceServerMiddleware
+ */
+class ResourceServerMiddleware extends Psr7ResourceServerMiddleware
 {
-    /**
-     * @var ResourceServer
-     */
-    private $server;
-
-    /**
-     * @param ResourceServer $server
-     */
-    public function __construct(ResourceServer $server)
-    {
-        $this->server = $server;
-    }
-
-    /**
-     * @param ServerRequestInterface $request
-     * @param ResponseInterface      $response
-     * @param callable               $next
-     *
-     * @return \Psr\Http\Message\ResponseInterface
-     */
-    public function __invoke(ServerRequestInterface $request, ResponseInterface $response, callable $next)
-    {
-        try {
-            $request = $this->server->validateAuthenticatedRequest($request);
-        } catch (OAuthServerException $exception) {
-            return $exception->generateHttpResponse($response);
-            // @codeCoverageIgnoreStart
-        } catch (\Exception $exception) {
-            return (new OAuthServerException($exception->getMessage(), 0, 'unknown_error', 500))
-                ->generateHttpResponse($response);
-            // @codeCoverageIgnoreEnd
-        }
-
-        // Pass the request and response on to the next responder in the chain
-        return $next($request, $response);
-    }
 }
